Updated to be upgrade if already installed

+++ run.sh
@@ -2,6 +2,10 @@
 
 PKG_PATH=$2
 
-sudo cp $PKG_PATH/scripts/install-chocolatey.ps1 /mnt/c/install-chocolatey.ps1
-powershell.exe -File C:\\install-chocolatey.ps1
-sudo rm /mnt/c/install-chocolatey.ps1
+if ! command -v choco.exe >/dev/null 2>&1; then
+  sudo cp $PKG_PATH/scripts/install-chocolatey.ps1 /mnt/c/install-chocolatey.ps1
+  powershell.exe -File C:\\install-chocolatey.ps1
+  sudo rm /mnt/c/install-chocolatey.ps1
+else
+  choco.exe upgrade chocolatey -y
+fi
